[Feature #163587283] Create a political party
